+++ docs/development/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### New features
 
+- [PR 49](https://github.com/OpenSenseAction/poligrain/pull/49) Add
+  point-to_point nearest neighbor lookup (by
+  [@cchwala](https://github.com/cchwala))
 - [PR 24](https://github.com/OpenSenseAction/poligrain/pull/24) Allow to color
   CML paths when plotting on map using a cmap (by
   [@cchwala](https://github.com/cchwala))